moved exceeding log test tosender  base class
