+++ logzio-sender/src/test/java/io/logz/sender/InMemoryQueueTest.java
@@ -5,10 +5,7 @@ import io.logz.sender.LogzioSender.Builder;
 import io.logz.sender.exceptions.LogzioParameterErrorException;
 import org.junit.Test;
 
-import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 
@@ -16,9 +13,6 @@ import static io.logz.sender.LogzioTestSenderUtil.createJsonMessage;
 
 public class InMemoryQueueTest extends LogzioSenderTest {
     private final static long defaultCapacityInBytes = 100 * 1024 * 1024;
-    private final static int MAX_LOG_LINE_SIZE_IN_BYTES = 32700;
-    private final static String TRUNCATED_MESSAGE_SUFFIX = "...truncated";
-    private final static String EXCEEDING_MESSAGE_FILE_PATH = "src/test/resources/exceeding_max_size_message.log";
 
     @Override
     protected Builder getLogzioSenderBuilder(String token, String type, Integer drainTimeout,
@@ -121,108 +115,6 @@ public class InMemoryQueueTest extends LogzioSenderTest {
         tasks.shutdownNow();
     }
 
-    @Test
-    public void checkExceedingMaxSizeJsonLogWithCut() throws LogzioParameterErrorException, IOException {
-        String token = "checkExceedingMaxSizeJsonLogWithCut";
-        String type = random(8);
-        String loggerName = "checkExceedingMaxSizeJsonLogWithCutName";
-        int drainTimeout = 2;
 
-        String message = Files.readString(Path.of(EXCEEDING_MESSAGE_FILE_PATH));
-        JsonObject log = createJsonMessage(loggerName, message);
-
-        int logSize = log.toString().getBytes(StandardCharsets.UTF_8).length;
-        ScheduledExecutorService tasks = Executors.newScheduledThreadPool(1);
-
-        Builder testSenderBuilder = getLogzioSenderBuilder(token, type, drainTimeout, 10 * 1000,
-                10 * 1000, tasks, false);
-        testSenderBuilder.setExceedMaxSizeAction("cut");
-        setCapacityInBytes(testSenderBuilder, logSize);
-
-        LogzioSender testSender = createLogzioSender(testSenderBuilder);
-
-        testSender.send(log);
-        sleepSeconds(2 * drainTimeout);
-        mockListener.assertLogReceivedByMessage(message.substring(0, MAX_LOG_LINE_SIZE_IN_BYTES - TRUNCATED_MESSAGE_SUFFIX.length()) + TRUNCATED_MESSAGE_SUFFIX);
-        tasks.shutdownNow();
-    }
-
-    @Test
-    public void checkExceedingMaxSizeBytesLogWithCut() throws LogzioParameterErrorException, IOException {
-        String token = "checkExceedingMaxSizeBytesLogWithCut";
-        String type = random(8);
-        String loggerName = "checkExceedingMaxSizeBytesLogWithCutName";
-        int drainTimeout = 2;
-
-        String message = Files.readString(Path.of(EXCEEDING_MESSAGE_FILE_PATH));
-        JsonObject log = createJsonMessage(loggerName, message);
-
-        int logSize = log.toString().getBytes(StandardCharsets.UTF_8).length;
-        ScheduledExecutorService tasks = Executors.newScheduledThreadPool(1);
-
-        Builder testSenderBuilder = getLogzioSenderBuilder(token, type, drainTimeout, 10 * 1000,
-                10 * 1000, tasks, false);
-        testSenderBuilder.setExceedMaxSizeAction("cut");
-        setCapacityInBytes(testSenderBuilder, logSize);
-
-        LogzioSender testSender = createLogzioSender(testSenderBuilder);
-
-        testSender.send(log.toString().getBytes(StandardCharsets.UTF_8));
-        sleepSeconds(2 * drainTimeout);
-        mockListener.assertLogReceivedByMessage(message.substring(0, MAX_LOG_LINE_SIZE_IN_BYTES - TRUNCATED_MESSAGE_SUFFIX.length()) + TRUNCATED_MESSAGE_SUFFIX);
-        tasks.shutdownNow();
-    }
-
-    @Test
-    public void checkExceedingMaxSizeJsonLogWithDrop() throws LogzioParameterErrorException, IOException {
-        String token = "checkExceedingMaxSizeJsonLogWithDrop";
-        String type = random(8);
-        String loggerName = "checkExceedingMaxSizeJsonLogWithDropName";
-        int drainTimeout = 2;
-
-        String message = Files.readString(Path.of(EXCEEDING_MESSAGE_FILE_PATH));
-        JsonObject log = createJsonMessage(loggerName, message);
-
-        int logSize = log.toString().getBytes(StandardCharsets.UTF_8).length;
-        ScheduledExecutorService tasks = Executors.newScheduledThreadPool(1);
-
-        Builder testSenderBuilder = getLogzioSenderBuilder(token, type, drainTimeout, 10 * 1000,
-                10 * 1000, tasks, false);
-        testSenderBuilder.setExceedMaxSizeAction("drop");
-        setCapacityInBytes(testSenderBuilder, logSize);
-
-        LogzioSender testSender = createLogzioSender(testSenderBuilder);
-
-        testSender.send(log);
-        sleepSeconds(2 * drainTimeout);
-        mockListener.assertNumberOfReceivedMsgs(0);
-        tasks.shutdownNow();
-    }
-
-    @Test
-    public void checkExceedingMaxSizeBytesLogWithDrop() throws LogzioParameterErrorException, IOException {
-        String token = "checkExceedingMaxSizeBytesLogWithDrop";
-        String type = random(8);
-        String loggerName = "checkExceedingMaxSizeBytesLogWithDropName";
-        int drainTimeout = 2;
-
-        String message = Files.readString(Path.of(EXCEEDING_MESSAGE_FILE_PATH));
-        JsonObject log = createJsonMessage(loggerName, message);
-
-        int logSize = log.toString().getBytes(StandardCharsets.UTF_8).length;
-        ScheduledExecutorService tasks = Executors.newScheduledThreadPool(1);
-
-        Builder testSenderBuilder = getLogzioSenderBuilder(token, type, drainTimeout, 10 * 1000,
-                10 * 1000, tasks, false);
-        testSenderBuilder.setExceedMaxSizeAction("drop");
-        setCapacityInBytes(testSenderBuilder, logSize);
-
-        LogzioSender testSender = createLogzioSender(testSenderBuilder);
-
-        testSender.send(log.toString().getBytes(StandardCharsets.UTF_8));
-        sleepSeconds(2 * drainTimeout);
-        mockListener.assertNumberOfReceivedMsgs(0);
-        tasks.shutdownNow();
-    }
 }
